+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 5.2.0
 -----------------
 * [Feature] 5.2 compatible version
+  * IMPORTANT: Since Moodle removed mimetex in 5.2, latex expressions will only work if the tex filter works. You can test it at https//your.moodle.instance/filter/tex/texdebug.php
 * [Bug] Fixed the look of the controls on the result overview
 * [Bug] Fixed that the order of questions wasn't correct in the single group view in the statistics overview
 * [Bug] Fixed that <code> end <pre> tags had a scroll bar [github #289]